# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/sms/list": {
            "get": {
                "description": "Returns the list of all sent SMS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SMS"
                ],
                "summary": "Get all SMS",
                "responses": {
                    "200": {
                        "description": "List of sent SMS",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SmsResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sms/send": {
            "post": {
                "description": "Sends an SMS to the specified phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SMS"
                ],
                "summary": "Send an SMS",
                "parameters": [
                    {
                        "description": "SMS request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SmsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sent SMS details",
                        "schema": {
                            "$ref": "#/definitions/model.SmsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sms/status": {
            "post": {
                "description": "Returns the status of an SMS by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SMS"
                ],
                "summary": "Check SMS status",
                "parameters": [
                    {
                        "description": "SMS ID payload",
                        "name": "smsId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SmsId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SMS status details",
                        "schema": {
                            "$ref": "#/definitions/model.SmsStatus"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SmsId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "model.SmsRequest": {
            "type": "object",
            "properties": {
                "certificate": {
                    "type": "string"
                },
                "client_token": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "model.SmsResponse": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "date_created": {
                    "type": "integer"
                },
                "date_sent": {
                    "type": "integer"
                },
                "extend_status": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "raw": {},
                "service_id": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "model.SmsStatus": {
            "type": "object",
            "properties": {
                "extend_status": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}